maintain: move access keys tests
